+++ js/main.mjs
@@ -1,0 +1,6 @@
+import './util.mjs';
+import {createFotos} from './data.mjs';
+import './form.mjs';
+
+// Временный вызов функций, чтобы ESLint не выводил ошибки о неиспользуемой функции
+createFotos();
